Rename aux to extra to make Estrella source code able to exist on MS Windows. Closes #25
